+++ src/main/java/com/teamo/teamo/repository/MemberRepository.java
@@ -3,6 +3,10 @@ package com.teamo.teamo.repository;
 import com.teamo.teamo.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface MemberRepository extends JpaRepository<Member, Long> {
     boolean existsByEmail(String email);
+
+    Optional<Member> findByEmail(String email);
 }
