+++ src/main/java/com/teamo/teamo/type/AuthType.java
@@ -1,5 +1,13 @@
 package com.teamo.teamo.type;
 
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
 public enum AuthType {
-    ROLE_USER, ROLE_ADMIN
+    ROLE_USER("USER"),
+    ROLE_ADMIN("ADMIN");
+
+    private final String key;
 }
